requirements.txt
# ============================================================
# StreamSpatial-VLM 依赖清单
# 硬件要求：NVIDIA RTX 3060 12GB，CUDA 11.8+
# Python 版本：3.10+
# ============================================================

# ---------- 深度学习核心 ----------
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ---------- Transformers 生态 ----------
transformers>=4.40.0
accelerate>=0.28.0
peft>=0.10.0

# ---------- Qwen2.5-VL 依赖 ----------
qwen-vl-utils>=0.0.8
tiktoken>=0.6.0

# ---------- VGGT 依赖 ----------
# vggt  # 从 GitHub 安装：pip install git+https://github.com/facebookresearch/vggt.git
einops>=0.7.0

# ---------- 深度估计 ----------
# depth-anything-v2  # 从 GitHub 安装
opencv-python>=4.9.0

# ---------- 数据处理 ----------
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0
h5py>=3.9.0

# ---------- 评估指标 ----------
nltk>=3.8.0
rouge-score>=0.1.2
pycocoevalcap>=1.2

# ---------- 配置与工具 ----------
pyyaml>=6.0
tqdm>=4.66.0
rich>=13.0.0

# ---------- 可视化 ----------
matplotlib>=3.7.0
seaborn>=0.13.0

# ---------- 性能分析 ----------
psutil>=5.9.0
py-cpuinfo>=9.0.0

# ---------- 开发工具 ----------
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
